UserDTO 패키지 변경 및 lombok 적용
